Refactor deeplay/models/__init__.py, deeplay/models/backbones/resnet18.py, deeplay/models/discriminators/__init__.py, deeplay/models/discriminators/cyclegan.py, deeplay/models/discriminators/dcgan.py, deeplay/models/generators/__init__.py, deeplay/models/generators/dcgan.py
