+++ deeplay/models/__init__.py
@@ -2,3 +2,5 @@ from .recurrentmodel import RecurrentModel
 from .visiontransformer import ViT
 from .backbones import *
 from .mlp import *
+from .generators import *
+from .discriminators import *
