+++ deeplay/models/discriminators/__init__.py
@@ -1,1 +1,2 @@
 from .cyclegan import CycleGANDiscriminator
+from .dcgan import DCGANDiscriminator
