+++ deeplay/models/generators/__init__.py
@@ -1,1 +1,2 @@
 from .cyclegan import CycleGANResnetGenerator
+from .dcgan import DCGANGenerator
