+++ deeplay/models/generators/dcgan.py
@@ -1,6 +1,6 @@
 import torch
 import torch.nn as nn
-from deeplay.blocks.conv2d.conv import Conv2dBlock
+from deeplay.blocks.conv.conv2d import Conv2dBlock
 from deeplay.components import ConvolutionalDecoder2d
 from deeplay.external.layer import Layer
 from deeplay.initializers.normal import Normal
@@ -86,7 +86,11 @@ class DCGANGenerator(ConvolutionalDecoder2d):
         class_conditioned_model: bool = False,
         embedding_dim: int = 100,
         num_classes: int = 10,
+        output_channels=None
     ):
+        if output_channels is not None:
+            out_channels = output_channels
+            
         self.latent_dim = latent_dim
         self.output_channels = out_channels
         self.class_conditioned_model = class_conditioned_model
